requirements.txt
# 核心框架
#torch==2.0.1+cu118
#torchvision==0.15.2

# 图像处理
opencv-python==4.7.0.72
Pillow==9.5.0
albumentations==1.3.0
scikit-image==0.20.0

# 数据处理
numpy==1.24.3
pandas==2.0.2
pyyaml==6.0  # 配置文件解析

# 训练工具
tensorboard==2.12.0
tqdm==4.65.0
codecarbon==2.2.4  # 碳排放跟踪

# 可视化
matplotlib==3.7.1
seaborn==0.12.2

# 工具类
tabulate==0.9.0  # 美观表格输出
hydra-core==1.3.2  # 高级配置管理（可选）